requirements.txt
pyqrcode
astroid==2.3.3
bson==0.5.8
colorama==0.4.3
dataclasses==0.6
Django==2.2.10
django-rest-framework==0.1.0
djangorestframework==3.11.0
djongo==1.2.38
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pylint==2.4.4
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
typed-ast==1.4.1
wrapt==1.11.2
